Migrate to bazel-deps
This allows us to store our maven dependencies in a nice yaml format,
rather than using lots of maven_jar invocations.

Unfortunately, it does mean having a separate bazel-deps checkout to
regenerate dependencies, but this is not a huge deal (and the author is
likely to make a release jar so we can run this internally).