Refactor settings and their application
These are now separate packages, one for the YAML layout and the other
for the application, which affords a clean separation of
responsibilities (and dependent libraries).

Each library has its own internal hierarchy mirroring its setup in
IntelliJ. This allows us to avoid namespacing conflicts (as it is
possible to have multiple JavaSettings for several settings).

A dependency-injection framework will be needed soon, as constructing
the configuration classes will get very cumbersome.