Avoid using System.exit(0) to terminate
After generating our IDEA files, IDEA would end up hanging. To work
around this, we shoved in a System.exit(0) but `bazel test` does not
allow this. As such, we use ApplicationEx.exit() to do this for us, and
will hopefully allow us to run bazel tests in the future.