Extract tests into "framework" using Junit suites
This will allow us to drop in a class and have bazel test this for us.
The problem is that we're just running one target, with no hint of
progress until everything is completed. Luckily, we still get separate
data in our test cases.

In future, we'll probably run with individual java_test invocations,
albeit using a custom rule to alleviate the boilerplate needed.