Convert idea-settings to an IntelliJ plugin
This will make everything far more stable as we try to upgrade to new
IntelliJ versions, and will make tests run faster as we can leverage
Jetbrains' framework.

We convert all integration tests to unit tests, and we'll investigate
using a proper integration testing framework in future to further
coverage.

Also update to 2019.2.