Replace bazel_deps with rules_jvm_external
This is the more 'official' way to manage maven dependencies, and
doesn't rely on us having the bazel-deps repo checked out to update
dependencies; we can do so entirely with our checkout.

Note that we tried to use dagger via bazel directly here but it didn't
take; we leave the dependencies here as they will otherwise be unused
until we take another stab at managing it.