Add basic support for inspections
This sets up the groundwork for adding future inspections; adding them
should come down to specifying a new class inheriting from a base
applier, and binding it in.

If the inspection has options, we will need to make another applier and
bind it in a similar manner in a subcomponent.