Add support for "Unnecessary call to 'super()'"
This is to battle-test adding extra inspections; it turns out if we make
the inspections without options use `Unit`, we will get duplicate
bindings. As a result, we need to declare a blank 'Options' for each
inspection, which can be just an object if none are available.

We should look into using class keys again, otherwise we'll have a lot
of extra declared options appliers in our InspectionsSubcomponent.